@@ -5,12 +5,6 @@ import { Observable } from 'rxjs/Observable';
 @Injectable()
 export class AppService {
 
-constructor(private http: HttpClient) { }
-
-getNpcDrops() {
-    this.http.get('localhost:8081/scrape/abyssal demon').subscribe(result => {
-
-    });
-}
+constructor() { }
 
 }
